We weren't using generic views in the dashboard because there's a bug in Q objects that was preventing us from creating a query for the "all incoming reviews" view.  We hacked around this by doing the OR+distinct step ourselves using lists.
It turns out that with just a little coaxing, django's generic
object_list view can be used with a list instead of a QuerySet.
This change does that, and makes the dashboard into a simple
superset of object_list.  The major benefit of this is that we
get pagination basically for free.

Reviewed at http://reviews.review-board.org/r/47/