This is the first step towards making Review Board easy to install and run for everyone.
I've decided to go with autotools here instead of setuptools, since it turns out
that setuptools is a ridiculously limited piece of garbage which falls flat on
its face as soon as you try to do something that isn't distributing a python
module.  Just try making an empty directory, or pre-processing a text file with
the prefix.  I mean it.  Garbage!

On the other hand, since autotools is actually designed to cope with the task of
building and distributing software, it makes everything easy.

There's more to do for this before everything is super sweet.
1. Verify that everything that needs to be installed is, actually, installed.
2. Make sure everything that we want to distribute is, actually, distributed.
3. Process the prefix into the sample configs so users can just point their
favorite web server at the configuration of their choice.
4. Write init scripts for various linuxes (although maybe just sysvinit is OK,
since I bet gentoo will write their own).
5. Build a virtual appliance.  Use this experience to make sure that there are a
minimum of steps for a user.
6. Update the GettingStarted wiki page.
7. Write other docs, such as README, AUTHORS, etc.
8. 1.0

Reviewed at http://reviews.review-board.org/r/193/