Link comments in review e-mails to the review page instead of the diff
This is something that's been irritating a lot of people, including me.  When
one gets a review e-mail from Review Board, it has the comments with links in
it.  These links have traditionally pointed to the relevant lines in the diff
viewer.  This is almost never what people want -- they want to look at the
review in the context of the source, and respond.

It makes a lot more sense to link to the relevant spot on the review page, from
which people can get to the diff if they want.  This change does exactly that,
adding anchors to the review so we can link directly.

Reviewed at http://reviews.review-board.org/r/214/
Fixes bug 322