Patch from Dan Sheridan to escape Subversion URLs.
Details:
This patch escapes the URL passed to Subversion so that paths with spaces in
them work correctly. Without this, a cryptic "Malformed repository URL" message
is shown to the user when submitting via the new review request form.

Testing done:
Verified working with patches to files with spaces in the filenames and spaces
in the paths, for both http:// and file:/// repository URLs.

Reviewed at http://reviews.review-board.org/r/361/