Patch from Vaclav Slavik to allow entering bug numbers as "#<num>" and not only "<num>"
When creating a review, the "Bugs" field accepts comma-separated list of bug
numbers. I've noticed that when you enter the number as e.g. #42, RB will
silently accept it, but create invalid URL with "#" included, because it expects
the bug numbers to be, well, numbers. But some people are in the habit of
entering bug references as #num (such as my colleague who entered it like that
into RB couple of days ago), so it would be nice if RB accepted this form too.

This patch does exactly that: when setting bug numbers, leading # is (silently)
removed. As a side effect, I also changed the way 'bugs_closed' field is
processed to be more robust: it recognizes not only '' as empty, but any
whitespace, and is now resistant to typos like "42,,43,".

Testing done:
Tested various combinations of inputs both with and without # and with/without
excessive use of "," and whitespace on our test installation of RB.

Reviewed at http://reviews.review-board.org/r/407/