Add initial support for monotone.
This is a patch from Gary Kramlich to add basic support for monotone. This
basically works.  It depends on a local database which we somehow need to
determine how to update.

Aside from that, binary files are not currently marked, and empty files cause
the diff viewer to blow up.

I've neglected to add it to the initial_data fixtures because it's in such a
preliminary state, but it's been sitting in our reviewboard for a while without
any progress; maybe having it here will encourage people to finish it up.

Reviewed at http://reviews.review-board.org/r/393/