If the summary is more than 300 chars Trim it down
Patch from JSlominski to fix bugs 438 and 515. When using post-review with
perforce there is an error if the CL description has more that 300 chars without
a newline. A SQL warning is generated (truncating summary) that will cause
post-review to throw an error 500. There is an example of the error page in Bug
515. This adds a small utility function to truncate the text down to 300
characters. If there are sentences in the summary, it will remove any hanging
sentence fragment at the end.

This is safe to do because it mostly crops up in the perforce client, which has
a copy of the summary as the first line in the description, so no data will be
lost.

Testing done:
I submitted and updated a review that had more than 300 chars without a \n and a
'.'.
I submitted and updated a review that had more than 300 chars without a \n and
did not have a '.'
I submitted and updated a review that had more than 300 chars without a \n and
had '.' after the 300th char

Fixes bug 438
Fixes bug 515

Reviewed at http://reviews.review-board.org/r/482/