Use CC field of email for group members on review requests.
This patch comes from Tarun Anupoju. When review groups do not have associated
mailing lists, we were adding the group members to the To: field. This is okay,
except it makes it a little harder for people to tell what's truly "to them" and
what's to the group in general when they're looking at an e-mail.

This change adds the following properties:
- In the emails that are sent out by Review Board, if there is a reviewer
  present( i.e someone mentioned in the people category) then the 'to' field
  will have the email of the Reviewer and the 'cc' field will consists of the
  other members present in the group(except the reviewer ...of course).
- If no one is mentioned then the email will be sent out normally.

Testing done:
Tested on local ReviewBoard setup (Fedora 9)
Ran unit tests (required some updates to the tests)

Reviewed at http://reviews.review-board.org/r/478/