Fix counting of diffs with drafts.
When a user has a draft diff, we counted wrong when choosing whether to show the
revision selector. We were not counting draft diffs in this total. The result of
this was:
- If you had only a draft diff, we showed a useless revision selector.
- If you have one published diff and one draft diff, we did not show the
  (potentially very useful) revision selector.
- If you had more than one published diff, it would show a revision selector.

Testing done:
Looked at the diff viewer with various combinations of published and unpublished
diffs.

Fixes bug 469
Reviewed at XXX