Add description about SITE_ROOT in settings_locals.py.tmpl
Patch from Benoît Allard to add a commented-out SITE_ROOT to the
settings_local.py template.

It tooks me half a day to discover that really useful line in the settings.py,
simply because I understood that file was for internal Django config, and that
no-one wanted to modify it. I finally felt on Christian's blog explaining that
setting and then read it in the settings.py file.

Such a setting should be accessible to people configuring their install of
reviewboard.

Reviewed at http://reviews.review-board.org/r/556/