Fix view diffs of files added to a branch (but not the trunk) in CVS
Patch from Matt McClellan to fix view diffs of files added to a branch (but not
the trunk) in CVS.

When I attempt to view the diff that was created from a file added to a branch
in CVS, I get a FileNotFoundError.

The problem is that files added to a branch (and not the trunk) are put in the
'attic' in the repository, which shows up as part of the pathname in the patch.
For example, if file foo.c is added to the frobnitz module on the bobo branch,
the RCS file entry in path in the patch file will look like this:

RCS file: /cvs_root_dir/frobnitz/Attic/foo.c,v

CVSClient is cvs.py is stripping the repository directory and the ','v suffix,
but leaving the '/Attic/' component of the path, which is what causes the
FileNotFound error.

Fixes bug 542.