Return an error if trying to set a summary with newlines.
At the moment, it's possible to create a review request that has a summary with
a newline in it. This works okay until RB tries to send email with that summary
as a subject, at which point it will blow up with an ugly backtrace.

I considered normalizing, but that's uglier. Since this is something that can
only be done via the JSON API, we might as well have the API return an error
instead.

Fixes bug 548
Reviewed at http://reviews.review-board.org/r/569/