Add column 'repository' containing the name of the repository in review datagrid
Patch from LAB Valentin to add column 'repository' containing the name of the
repository in review datagrid. As DVCS are gaining popularity, projects tends to
divide to small distinct repositories (which is partly due to the incapacity of
the main DVCS to retrieve sub-directories of a given repository).

As a matter of fact, code-review tools must be able to span accross several
repositories, which is the case of the reviewboard.  But a simple feature seems
missing : in the global review table there is no column to pin point to which
repository a particular code-review applies.

Fortunately, this do not take long to patch.

I'm sorry if this was previously discussed or even resolved. I'm using the last
public SVN code source, and this issue doesn't seem to have been worked out.

Reviewed at http://reviews.review-board.org/r/674/