Show a pretty error when attempting to create a review from an empty changeset.
Fairly frequently (at least at VMware), people try to create a review request
from the web UI by entering a changeset number that has no associated open
files. In these cases, we get a KeyError for 'depotFile' when attempting to get
the change description from the Perforce Python API.

This change adds a check and an exception to show an error on the field instead
of just sending a traceback to my email and dropping them back on the form.

Testing done: ran scmtools unit tests

Fixes bug 706.