Improvements to progressing loading of diff chunks.
The page no longer jumps around due to the size of diff chunks. We now build a table of roughly the right height and put a spinner in the middle. In most cases (where lines aren't wrapping), the resulting size will be the same. The general effect is that loading of diff fragments feels a lot smoother, and anchors are far more reliable.