Fix a race condition when publishing reviews.
We were attempting to set up an async chain for saving each individual comment, but instead of listening for when the save completed, we were just listening for the inlineEditor's "completed" signal, which just fires when the editor saved back to the <pre> tag. This is the same signal that our code to actually save to the server listens to.

The solution is to have a "saved" signal that is emitted when we know the save went through.

This should fix bug #820.