Clearcase module with post-review support
Patch from Daniel Belz. This is a clearcase integration module with post-review
support. It supports windows, linux and cygwin environments, and both snapshot
and dynamic views.

The module depends on a mounted view on the reviewboard server. It is capable of
submit reviews for checked out and checked in elements.

Testing done: Submitted reviews on linux, windows and cygwin using snapshot and
dynamic views.

Reviewed at http://reviews.review-board.org/r/653/