This is a much-needed rewrite of the Review Board website.
This is both a design and architectural rewrite. Instead of using PHP,
we're now using Django as the backend. This allows us to better
integrate parts of the site, such as news, with other parts. Down the
road, we'll be able to add such things as an extension subsection of the
site, an inline doc browser, and more.