Lots of optimizations for the diff viewer. We were calling the highlightChunk's updatePosition code way too frequently. Every time there was a key press in the comment box, we'd update the text status, causing a DOMNodeInserted event to fire, causing us to re-highlight the currently selected region.
We're now a lot more careful about only setting the text status the first time the comment box becomes "dirty," and we also have a lot of checks for updatePosition to only do work if this is an event we care about and if the position did in fact change. We also limit the number of computations we do in that function, since we were performing the same computations over and over again.

This should fix many of the major slowdown issues noticed in the diff viewer.