Add an activity indicator for AJAX calls.
This activity indicator is triggered during AJAX operations such as
fetching the review form, marking a change as submitted, saving
comments, or saving parts of review requests. This appears in the
top-right of the page in a simple but eye catching little box,
regardless of where you are on the page.

Reviewed at http://reviews.review-board.org/r/798/
Fixes bug #945