Change the Ship It column to show the number of Ship It reviews.
This change adds the number of reviews marked Ship It to the existing
Ship It column. Instead of showing just a pre-made icon, we now have a
little stylized box with a checkmark and the number beside it.

This ends up looking pretty decent on browsers that suppor rounded
corners. On IE, this will just be a box, but that's not too horrible.

Reviewed at http://reviews.review-board.org/r/838/