Make the page load for the review request page less jumpy.
When displaying a draft banner for either the review request or a
review, the page was a bit jumpy. We'd render the page and then display
the draft banners.

This change shows the draft banners right away, and even adds a nice
subtle slide in/out animation for the review draft banner at the top of the
window.

Reviewed at http://reviews.review-board.org/r/850/