Fix a showstopper issue where rb-site would fail on new installs.
This is caused by our dependency checking, which looks for modules
by doing an __import__, but since we do this after populating sys.path
with our (yet to be created) conf directory, Python sees the
non-existant path and doesn't attempt to import from that directory
again. The solution is to set this path after we've created and
populated the conf directory.

Fixes bug #1012.