Satisfy PostgreSQL's GROUP BY requirements by instead wrapping all references to reviews_review.timestamp with a MAX().
GROUP BY, as recommended in the PostgreSQL output, won't work for this
query, but PostgreSQL only requires the GROUP BY if we reference an
un-SELECTed column outside of an aggregate. So MAX() provides us a
simple way of keeping that column inside an aggregate, and shouldn't
introduce any speed issues since we limit this to 1 row.

Fixes bug #1099.